+++ pages/api/password.ts
@@ -10,6 +10,7 @@ import { ApiError } from 'next/dist/server/api-utils';
 import { recordMetric } from '@/lib/metrics';
 import { getCookie } from 'cookies-next';
 import { sessionTokenCookieName } from '@/lib/nextAuth';
+import env from '@/lib/env';
 
 export default async function handler(
   req: NextApiRequest,
@@ -60,16 +61,18 @@ const handlePUT = async (req: NextApiRequest, res: NextApiResponse) => {
   });
 
   // Remove all sessions other than the current one
-  const sessionToken = getCookie(sessionTokenCookieName, { req, res });
+  if (env.nextAuth.sessionStrategy === 'database') {
+    const sessionToken = getCookie(sessionTokenCookieName, { req, res });
 
-  await prisma.session.deleteMany({
-    where: {
-      userId: session?.user.id,
-      NOT: {
-        sessionToken,
+    await prisma.session.deleteMany({
+      where: {
+        userId: session?.user.id,
+        NOT: {
+          sessionToken,
+        },
       },
-    },
-  });
+    });
+  }
 
   recordMetric('user.password.updated');
